+++ sonar-project.properties
@@ -4,3 +4,11 @@ sonar.sources=src
 sonar.tests=tests
 sonar.sourceEncoding=UTF-8
 sonar.rust.lcov.reportPaths=lcov.info
+
+# Test code lives in src/ alongside production code (sibling *_tests.rs files
+# linked via `#[cfg(test)] #[path = "X_tests.rs"] mod tests;`). Exclude them
+# from copy-paste detection — test-fixture boilerplate is categorically
+# different from production duplication, and forcing extraction would create
+# bad abstractions per the project's discipline rule (see
+# docs/specs/2026-04-27-sonar-refactor-design.md).
+sonar.cpd.exclusions=**/*_tests.rs
